Change database option name, rename test case
