+++ tests/ORM/Functional/Driver/Common/Integration/Case383/Entity/Comment.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case5\Entity;
+namespace Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case383\Entity;
 
 use DateTimeImmutable;
