+++ tests/ORM/Functional/Driver/Common/Integration/Case383/schema.php
@@ -7,11 +7,11 @@ use Cycle\ORM\Relation;
 use Cycle\ORM\SchemaInterface as Schema;
 use Cycle\ORM\Select\Repository;
 use Cycle\ORM\Select\Source;
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case5\Entity\Comment;
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case5\Entity\Post;
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case5\Entity\PostTag;
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case5\Entity\Tag;
-use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case5\Entity\User;
+use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case383\Entity\Comment;
+use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case383\Entity\Post;
+use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case383\Entity\PostTag;
+use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case383\Entity\Tag;
+use Cycle\ORM\Tests\Functional\Driver\Common\Integration\Case383\Entity\User;
 
 return [
     'comment' => [
